refactor: move compute_normal_values function to compute_normal_values_center.py script
